backend/requirements.txt
geopy
google-api-python-client
google-auth
google-generativeai
grpcio
huggingface-hub
httpx
Jinja2
keras
monai
nibabel
numpy
opencv-python
pydantic
pydicom
requests
scikit-learn
scipy
tensorboard
tensorflow
timm
torch
torchvision
tqdm
uvicorn
python-dotenv
fastapi
fastapi-users
pytesseract
pymupdf
Pillow